Fix #66, remove compiler extensions from tbldefs
Do not declare tables as "static" w/attribute "used", neither are
needed, and build works fine without any special sauce.

This also makes the table name consistent by removing the "Default" in
the name - it is still the same table, the name does not need to
change just because it happens to be the default value.