Fix #98, use "GENERATED_FILE" option to generate_config_includefile
For an EDS build, the include files are generated, and do not yet exist
when CMake runs.  The generate script now has more error checking than
it used to and now throws an error if the file does not exist.

Use of the GENERATED_FILE flag causes it to bypass the the search/verify
and unconditionally link to the file.